destroy training queue jobs on takeover only if queue is not frozen

+++ app/models/settlement/settlement.rb
@@ -202,16 +202,14 @@ class Settlement::Settlement < ActiveRecord::Base
     
     # destroy all construction jobs
     self.queues.each do |queue|
-      queue.reload
-      queue.jobs.reload
       queue.jobs.destroy_all          unless queue.jobs.nil? # will remove also active job and event if existing
     end
 
     # destroy all training jobs
     self.training_queues.each do |queue|
-      queue.reload
-      queue.jobs.reload
-      queue.jobs.destroy_all          unless queue.jobs.nil? # will remove also active job and event if existing
+      unless queue.frozen?
+        queue.jobs.destroy_all          unless queue.jobs.nil? # will remove also active job and event if existing
+      end
     end
 
     self.garrison_army.destroy        unless self.garrison_army.nil?
